od: crash fix while importing a SEGY file and some error occurred.
Summary:
bug fix for 2408
http://bugs.dgbes.com/view.php?id=2408
uiMSG() sometimes take uisegyexamine window as a parent for its messageBox and when object of uisegyexamine is deleted by a timer in uiobjdisposer it tries to delete uiMessageBox of uiMSG() and bam.
Reproducible only in windows.

Test Plan: will only be able to test after it goes into release version. Crash stops happening mysteriously on its own on my system.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2790

git-svn-id: https://svn.opendtect.org/od/trunk@32807 a985855d-d384-4b45-ae33-31a410998c3c