Removed potential deadlock by forcing write-work to different thread.
Summary:
SeisSequentialWriter::iterateBuffer submitted work to a single threaded queue. Under certain circumstances,
the queue decided that it wished to do the work immediately from the same thread. That lead to a deadlock as
the the mutex was locked a second time.

Test Plan: Not tested. Should be tested by reviewer after acceptance. Apart from that visually inspected.

Reviewers: aneesh

Reviewed By: aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D2794

git-svn-id: https://svn.opendtect.org/od/trunk@32811 a985855d-d384-4b45-ae33-31a410998c3c