Bugfix: prestack data display in 3D scene showed texture with uniformly colored haze over it.
Summary:
Prestack data display in 3D scene showed texture with uniformly colored
haze over it. The root material of the prestack data display should NOT
be set with the gridlines color, because then the texture will inherit it
as well. The gridline color corresponds with the marker color of the colormap
of the texture, which explains the particular color of the haze seen on top
of the texture.

Test Plan:
The texture of the prestack data display looks ok again after removing the
"setMaterial(gridlinematerial_)" statement, and so far it looks like its
removal does harm anything else.

Reviewers: ding

Reviewed By: ding

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2813

git-svn-id: https://svn.opendtect.org/od/trunk@32857 a985855d-d384-4b45-ae33-31a410998c3c