od4.6: Implementation of getting CBVS file's full path in 2D seismics improved.
Summary:
gtFileName() is not implemented properly but due to current settings it was working fine.
It lacks flexibility that will be needed for new 2D seismics data storage design.
The setPath() of FilePath only takes the last part of its content and adds it to the string passed as srgument.

Test Plan: Runs 4.6 against the new 2D seismics data storage design and it works fine

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D2823

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32868 a985855d-d384-4b45-ae33-31a410998c3c