Adapted OpendTect horizon and fault (vertex shape) texture functionality for viewers with redraw-on-demand policy.
Summary:
No longer allowed to have a continuous subscription to the update callback,
because this makes the redraw-on-demand policy of the viewer void. Apart from
dealing with an update callback instead of setting the number of children
requiring update traversal, the code changes to be made are analogous to
those of TexturePlane and PanelStrip in osgGeo (see Phabricator D2816).

Test Plan:
Did various textural user modifications on horizons and faults in OpendTect,
with and without shaders, to see whether it still yields an immediate update
in the 3D scene. Also put breakpoints in OSG to assure that no longer redraws
are done in absence of any user action with only these texture objects in the
scene.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2827

git-svn-id: https://svn.opendtect.org/od/trunk@32906 a985855d-d384-4b45-ae33-31a410998c3c