Bugfix: Object info in status bar was not displayed in view mode (MAIN)
Summary:
Object info in status bar was not displayed in view mode. Nodes do not have
a pickstyle as in Coin, but have to be made unpickable by blocking the
intersection traversal. The most flexible and transparent solution is to
distinguish between two intersection traversal masks, an active one for real
clicks and a passive one for just hovering to get status info for instance.
This requires two intersection traversals by the EventCatchHandler, because
users do not want to click accidentally on an object that is behind another
object on which only hovering is allowed.

Test Plan:
Status info is shown again in view mode, while picking a fault or a horizon
on an inline is indeed disabled. Also went along all setPickable statements
to decide which "false" ones should not allow passive intersection traversal
either.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2858

git-svn-id: https://svn.opendtect.org/od/trunk@32927 a985855d-d384-4b45-ae33-31a410998c3c