Bugfix: DepthTabPlaneDragger poorly reacted on mouse actions because of z-fighting with its TexturePlane (after Phabricator D2858)
Summary:
After the introduction of active and passive intersection traversal masks,
DepthTabPlaneDragger poorly reacted on mouse actions because of z-fighting
with its TexturePlane. It turns out that draggers need their own traversal
mask, because an active dragger wants to mask its own TexturePlane but yet
keep it passively pickable for displaying status info. At the same time it
should not mask other (actively or passively) pickable objects from the
scene, because these need to be checked for possibly occluding the dragger,
in which case the dragger will not be activated. (Note that PolygonOffset
applies to (OpenGL) displaying only, and is not taken into account by
intersection traversals.

Test Plan:
Draggers for PlaneDataDisplay and PreStackiDataDisplay react on mouse
actions all the time, while status info of the texture is still displayed.
Also moved another plane in front of it, and indeed it was is impossible
to activate the occluded dragger. BoxDragger traversal will be modified
analogously, but will be committed later together with other changes
in that area.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2866

git-svn-id: https://svn.opendtect.org/od/trunk@32948 a985855d-d384-4b45-ae33-31a410998c3c