Functionality to have mouse interaction of the (volume) box dragger depend on users settings.
Summary:
(1) Added functionality to have the mouse interaction of the box dragger for
volumes depend on user settings. New code analogous to the existing code for
DepthTabPlaneDragger and used by PlaneDataDisplay, but their settings do
not have to be equal (although they will by default). Implemented extra option
to interpret in-depth translation of the dragger box as a box resize event,
which is what users seem to expect when "traversing" through the current
volume.
(2) Bugfix: Volume box dragger was unable to translate or scale beyond current
size of texture cube after a last-minute update to Phabricator D2678, in which
box dragger was snapped to the survey in (what turned out) not the right way.
(3) Adapt BoxDragger to introduction of separate DraggerIntersecTraversalMask
(actually belongs to Phabricator D2866)

Test Plan:
(1) Verified that the original OSG mouse interaction for the in-depth and
in-plane translation of the dragger box has been overruled by the defaults
users know from the plane draggers. Still to be implemented is a dialog
analogous to the uiVisPlaneDataDisplayDragProp dialog, and next make the
mouse interaction settings of volumes editable via an extra tab in their
Properties menu item dialog.
(2) Volume box dragger can be moved beyond the size of the current texture
cube again, and yet aligns with the survey steps.
(3) Volume box dragger reacts on mouse actions all the time, no z-fighting
with the texture cube anymore.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2873

git-svn-id: https://svn.opendtect.org/od/trunk@32974 a985855d-d384-4b45-ae33-31a410998c3c