od: Fix for creating moc file with Qt5
Summary:
After the move of the Qt files from include to src, the moc file were not generated anymore. Solved it by using the same mechanism as with Qt4 and disabling the automoc.
Part of T96

Test Plan: Moc files are generated again and all modules build and link.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Maniphest Tasks: T96

Differential Revision: https://phabricator.dgbes.com:8443/D2895

git-svn-id: https://svn.opendtect.org/od/trunk@33001 a985855d-d384-4b45-ae33-31a410998c3c