od4.6: Pre-load seismic data on restore session
Summary: Data which has been preloaded when saving the session, will be pre-loaded again when the session is restored.

Test Plan:
Tested a couple of sessions, with and without pre-loaded data. All works as expected.
A future improvement is to check whether current pre-loaded data is also requested in the session.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D2903

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33026 a985855d-d384-4b45-ae33-31a410998c3c