Bugfix: Selected object gets deselected when switching to view mode and manipulating the scene.
Summary:
Selected object gets deselected when switching to view mode and manipulating
the scene. Reason is that view mode implementation differs from Coin by masking
the "active intersection" traversal. This makes the event catcher trigger the
nothandled notifier when clicking the scene, not only in act mode but also in
view mode.

Test Plan:
Add an inline, move its dragger to another position, click on view mode button,
translate or rotate the scene with the mouse, and now this no longer triggers
the inline to be moved and reloaded.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2961

git-svn-id: https://svn.opendtect.org/od/trunk@33095 a985855d-d384-4b45-ae33-31a410998c3c