od: Added a function to get GeomID in GeometryManager given linename and lineset name.
Summary:
Background: If different linesets has lines with common names then concatenation of lineset name and linename(with a '_' in between) will be used to form a unique linename.
This function will be needed for 2D data storage format conversion.

Test Plan: Using it in conversion class.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Maniphest Tasks: T104

Differential Revision: https://phabricator.dgbes.com:8443/D2968

git-svn-id: https://svn.opendtect.org/od/trunk@33110 a985855d-d384-4b45-ae33-31a410998c3c