Added support for touch-events in the 3D visualization.
Summary:
The idea is that we are taking the Pinch touch event from qt, and sending it into the osg-events.

Eventually, this will be done in osg itself, but we have to wait for changes in the trunk.

Test Plan:
The pinch gesture works fine, after some adaptions to osgGeo. These will come
i a separate revision.

Reviewers: jaap

Reviewed By: jaap

Maniphest Tasks: T98

Differential Revision: https://phabricator.dgbes.com:8443/D3007

git-svn-id: https://svn.opendtect.org/od/trunk@33159 a985855d-d384-4b45-ae33-31a410998c3c