od4.6: added protections in FFTFilter class when building the frequency taper window
Summary:
fix for solving a reported crash (Mantis #0002469) on line 158 of fftfilter.cc. The only thing I can guess
is that the lowpass frequency is larger than the high pass frequency. Thus I added protections in the non-UI but also UI parts

Test Plan: Error messages are issued, no crash when I disable the error reporting in the UI.

Reviewers: bert

Reviewed By: bert

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D3003

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33161 a985855d-d384-4b45-ae33-31a410998c3c