Made code chose osgSim colorbar when available.
Summary: The idea is to switch to osgSim::Scalarbar as soon as we could, and hence disable the osgGeo one as it won't be needed.

Test Plan: Works fine when compiling from trunk. Tested to change the macro (to also simulate osgGeo-usage), and that also works fine.

Reviewers: ranojay

Reviewed By: ranojay

Maniphest Tasks: T103

Differential Revision: https://phabricator.dgbes.com:8443/D3011

git-svn-id: https://svn.opendtect.org/od/trunk@33173 a985855d-d384-4b45-ae33-31a410998c3c