od: Make available alternative property in PropRefSelection
Summary:
see Mantis #2447: In the selection of properties, give the possibility that the alternative property should be selected by default.
Example: You request a P-wave input, but prefer to auto-select the Sonic logs instead of the P-wave logs from the database, because they are nearly always the preferred input.


Test Plan: tested on a well without setup, using the Synthetic-to-seismic tie manager, now the Sonic log is auto-selected instead of Vp log when both are present.

Reviewers: helene

Reviewed By: helene

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3010

git-svn-id: https://svn.opendtect.org/od/trunk@33175 a985855d-d384-4b45-ae33-31a410998c3c