Made DataPack aware of its manager (if any).
Summary:
I see numerous places where datapacks are used carelessly. One example are when
'observing' datapacks are obtained as local variables. There is no guarantee that the objects will continue to live, and hence they should be obtained with a non-observing state. The included copy is however not critical as it can be assumed things are handled singlethreaded there.

This would however lead to the problem that if one obtains a datapack locally, it has to be released when we are finished. If this is not done, we get a memory leak.

I am therefor proposing a DataPackReleaser that releases the datapack when it goes out of scope. To make it work, the releaser needs to know which manager manages it. I decided to put this information in the datapack itself, and give the datapack a relase() function.

Additionally, I fixed some inconsistencies in the DataPackMgr::release mechanism with regards to threading.

Test Plan: All tests work according to plan.

Reviewers: bert

Reviewed By: bert

CC: helene, mahant

Differential Revision: https://phabricator.dgbes.com:8443/D3017

git-svn-id: https://svn.opendtect.org/od/trunk@33178 a985855d-d384-4b45-ae33-31a410998c3c