od: Crash due to null pointer.
Summary: Crash when checkbox for using existing depth/time model is checked without selecting any well in the window for selecting data to tie well to seismic.

Test Plan: Crash is gone and things work as they would have if some well is already selected before the checkbox is checked.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3018

git-svn-id: https://svn.opendtect.org/od/trunk@33187 a985855d-d384-4b45-ae33-31a410998c3c