[od] Fix for crash report (Mantis ID 2503 and 2415).
Summary:
We have two similar crash reports 2503 and 2415.

http://bugs.dgbes.com/view.php?id=2503.
http://bugs.dgbes.com/view.php?id=2415. 

As per the call stack the most likely reason for crash is when ms is null.
So added check for the same.

Test Plan:
Only added null pointer check.
Did some very simple tests by adding and removing attribute but could not reproduce the crash.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3058

git-svn-id: https://svn.opendtect.org/od/trunk@33224 a985855d-d384-4b45-ae33-31a410998c3c