[od] Fixing problem in attribute selection on windows 32bit.
Summary:
This problem was reported in mantis bugs 2475 and 2449.
http://bugs.dgbes.com/view.php?id=2475.
http://bugs.dgbes.com/view.php?id=2449.

The problem here is the order in which different operators and functions are called. Windows 32 Bit compiler seems to differ from rest.

So moved these two function calls in a separate line, so as not to be affected by these compiler dependent behaviour.

Test Plan: Tested these changes on a 32bit machine.

Reviewers: raman

Reviewed By: raman

CC: kristofer, helene

Differential Revision: https://phabricator.dgbes.com:8443/D3070

git-svn-id: https://svn.opendtect.org/od/trunk@33246 a985855d-d384-4b45-ae33-31a410998c3c