od4.6: A reasonable fix for the well marker sorting problem (Mantis 2518)
Summary:
Not all wells contain all markers. When merged in a simple way things can get really out of hand.
The first problem is that the depths (dah values) of the two wells are simply incompatible. Thus, you have to set up a matching framework.
The next problem, even bigger, is that although you can rather easily merge 2 wells, the merged product may contain an ordering that conflicts with what you find in a 3rd well. Imagine two markers in the same zne, each know in one of the wells. Then without more info they can be placed in the wrong order. Then a 3rd well may have this info (which of the 2 comes first). The current algorithm therefore first uses the 3rd well to correct the ordering in the merged product before adding new markers.

Test Plan: Tried in Marathon_Eagleford_CentralMerge_2013, Fe_Demo and Penobscot. Seems to do a fine job.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D3087

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33270 a985855d-d384-4b45-ae33-31a410998c3c