Implementation of translation event traversals.
Summary:
The plan is that the uiBaseObject has a virtual function called translate() (will later be renamed to translateText()). All objects who wish to take action when the translation is changing should have an own implementation. One instance of that is hence the uiParent, which will trigger this on all children.

The idea is then that the uiMainWin will be notified when the language is changing (via callback), and then start the traversal of its tree.

Test Plan:
This is not tested. The idea now is to put the infrastructure in place in a way that does not jeopardizes stability in non-translation operations.

OpendTect starts fine.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3139

git-svn-id: https://svn.opendtect.org/od/trunk@33350 a985855d-d384-4b45-ae33-31a410998c3c