Made MenuHandler use uiString for text items, and adapted throughout.
Summary:
Also moved uiString to Basic, as it has to be available in General, but General is not directly dep on Qt, so I could not put it there.

Additionally, I adopted UserInput fld to take uiString as well for the tooltip.

Test Plan: OpendTect starts nicely, the menus generated by menuhandler works as expected. The tooltips works as expected.

Reviewers: nanne

Reviewed By: nanne

CC: helene

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3154

git-svn-id: https://svn.opendtect.org/od/trunk@33371 a985855d-d384-4b45-ae33-31a410998c3c