[od4.6] Potential fix for crash when uiSurveyMap::reDraw function is called.
Summary:
http://bugs.dgbes.com/view.php?id=2534
http://bugs.dgbes.com/view.php?id=2539

As per the call stack, the most likely reason for crash is when survinfo_ is already deleted. But uiSurveyMap still holds the pointer to deleted object.
So setting survinfo_ to null in uiSurveyMap whenever survinfo_ is being deleted or being set to null in uiSurvey.

Test Plan:
Tested a few combinations of survey change
1. .survey file is corrupt.
2. When user does not completely go through the process of setting new survey and presses cancel... and rollBackSurvey is called.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3160

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33378 a985855d-d384-4b45-ae33-31a410998c3c