Adapted combobox to uiStrings internally.
Summary:
I adapted the principle that the uiStrings are only expanded internally, and all returns should return nice ascii strings.
That should work well as long as the object is read-only. Then, we can always return the getFullString of an item as the text.

Test Plan: Comboboxes does still work.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3165

git-svn-id: https://svn.opendtect.org/od/trunk@33412 a985855d-d384-4b45-ae33-31a410998c3c