od: Crash due to deleted well data.
Summary: Mantis IDs: 2227, 2294. uiWellDahDisplay does not get notified if Well::Data is going to be deleted. Hence the change.

Test Plan: Added a well in the 3D scene. Opened the same well in tie well to seismics window. Removed the well from the 3D scene and moved mouse over the well in tie well to seismics window to reproduce the crash. The crash did not reappear. Also checked if all objects of Well::Data are properly deleted.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3176

git-svn-id: https://svn.opendtect.org/od/trunk@33422 a985855d-d384-4b45-ae33-31a410998c3c