[od] : Adding test program for comparing angle computation and raytracing in different surveys.
Summary:
Added test program for comparing angle computation in different surveys
The test covers surveys in time and depth( meter, feet ).

1. The test compares angles at different offsets and different depths( top, bottom, middle and another position as specified by Arnaud in the phabricator task ).

2.a) Also here assumption is made that offsets are always in meters. So for all surveys same offset has been defined. 

2.b) http://intranet/wiki/pmwiki.php/Main/DepthUnitPolicy. 
Following the depth policy, the depths are assumed to be in meter or feet depending on the survey. 

The result of the test program are similar, but not exactly same.

Test Plan: Tested by running test program. Also stepped through the code in the debugger.

Reviewers: kristofer, arnaud

Reviewed By: kristofer

CC: raman

Maniphest Tasks: T81

Differential Revision: https://phabricator.dgbes.com:8443/D3182

git-svn-id: https://svn.opendtect.org/od/trunk@33428 a985855d-d384-4b45-ae33-31a410998c3c