od: New Vint value in TD manager in not in the correct UoM, and errors during export to Ascii
Summary:
Two problems with well with "Display depths in feet":
Vint is always reported in internal units, not in the user unit. 
The exported time-depth curve as wrong TVDSD and TVDSS values, because a conversion factor was applied too soon.

Test Plan: Tested in client survey with "Display depths in feet"

Reviewers: salil

Reviewed By: salil

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3210

git-svn-id: https://svn.opendtect.org/od/trunk@33440 a985855d-d384-4b45-ae33-31a410998c3c