Two bugfixes for messy fault surfaces after importing faults (porting Phabricator issue D3232 to OD4.7)
Summary:
(1) Fault surfaces were twisted in-between some sticks. The untwistStick()
function did no longer work correctly after revision r25746 in August 2012.
Looks like its code has been modified to repair, but without identifying
the real problem. After the readSectionForImport(.) has been extended with
intelligent code to merge one-knotted or attaching sticks together, it is
no longer sound to call the isTwisted(.) function of the input faultstickset,
because there is no one-to-one relationship anymore between the sticks of
the input faultstickset and the sticks in the sticks_ object set, in which
the fault is constructed. Furthermore, no objection against using an angular
test to mark two sticks as twisted, but the criterion should be that the
cosine yielded by their dot product is negative.
(2) The merging of two sticks on the same section sometimes resulted in a
V-shaped stick because of connecting the wrong ends together. The picking
order of either stick was not taken into account.

Test Plan:
Tested with a big set of faults to be imported for the WINO-Kvalross_PL611
survey located at d52/surveys. All cases of twisted fault surfaces are
gone. No V-shaped sticks anymore when two sticks on the same section are
merged.

Reviewers: yuancheng

Reviewed By: yuancheng

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3238

git-svn-id: https://svn.opendtect.org/od/trunk@33489 a985855d-d384-4b45-ae33-31a410998c3c