od4.6: Fix for marker set sorting introduced in this patch release
Summary: If markers needed to be moved the algo would crash, due to an overloaded += operator. Now made sure that the names of markers that need to be moved are empty when adding them in another place (and removeing the old ones afterwards).

Test Plan: Talisman Leven survey well attrib extraction doesn;t crash anymore.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D3262

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33525 a985855d-d384-4b45-ae33-31a410998c3c