od4.6: Made sure lithologs are never interpolated
Summary: A protection was already in place in calcVal, but neither in getValue nor in gtVal against interpolation of litho-like logs. Since it needed to be centralised, and to gain some runtime speed, I added a class member iscode_. It is set by the reader, and of course editing one value may break it, but their is no log editor for now so it is relatively safe I believe (please confirm)

Test Plan: tested on Leven dataset. No more decimals outputs

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D3267

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33530 a985855d-d384-4b45-ae33-31a410998c3c