Coordinate change callbacker
Summary:
Right now, since we changed osg frame schema to ON_DEMAND, sometimes geometry is not refreshed. adding dirtyCoordinates or requestSingleDraw to all places where 
the coordinate changed is not good idea. So I added a callbacker notify in visCoord class, and it will be called in vertexshape or other classes ( if this idea is ok, 
later on I will can check other places in od such as markerset... ).

Test Plan: Compiles and Tested in Polyline fine

Reviewers: kristofer

Reviewed By: kristofer

Maniphest Tasks: T74

Differential Revision: https://phabricator.dgbes.com:8443/D3309

git-svn-id: https://svn.opendtect.org/od/trunk@33662 a985855d-d384-4b45-ae33-31a410998c3c