Call new OsgGeo functionality (Phabricator D3355) for minimizing texture color vagueness (Mantis 2542)
Summary:
1) Calling new OsgGeo functionality to minimize texture color vagueness.
The combination of mipmap filtering, color mapping and a fragment shader moves
the locally averaged texture color towards the middle value of the color map.
Solved by allocating extra texture channel(s) in which all values can be
encoded squared, from which the shader can derive the local standard deviation
of the texture values and use that as offset into the scale space of the color
map.
(2) Small rework of TextureChannels class, going from a hard-coded fixed number
of texture bands to an (internally) flexible number of texture bands. Note that
those "bands" are called "channels" in the osgGeo context, but channels already
correspond to attributes in this class, and may consist of several bands each.

Test Plan:
Tested for inlines and 2D lines, at different zoom distances, viewing angles
and depth stretch, for different color maps and for attributes with and without
undefined areas, with enabled and disabled shader. Great improvement in color
stability. Using only one byte to encode the squared values seems sufficient
for seismic textures! Let's wait for counter examples.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3356

git-svn-id: https://svn.opendtect.org/od/trunk@33696 a985855d-d384-4b45-ae33-31a410998c3c