Added implementation of translation tools, as well as tools for handling the application name.
Summary:
This finalizes the infrastructure part of the translation. There may still be some build-stage stuff to do, but that can wait.

The translations will sit in a directory under data, and based on directory-
listing, they will be made available.

I had to change a few things as we need to support multiple simultanious translation files, as od and dgb will be different from arkcls and other vendors.

Test Plan: Compiles and links. We don't have any translation-files to play with.

Reviewers: nanne

Reviewed By: nanne

CC: helene

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3332

git-svn-id: https://svn.opendtect.org/od/trunk@33702 a985855d-d384-4b45-ae33-31a410998c3c