od: Added program od_batch_launcher
Summary: For Unix platforms the program will launch a given process and if monitoring is needed, will also launch the od_Progress_Viewer program with the process ID of the batch program. Also supports output redirection to a temp file.

Test Plan:
Tested on Linux:
1. with od_process_attrib where the monitor file is the log file.
2. with od_glxinfo where the output is redirected to a temp file which is read by od_Progress_Viewer

Pending test on Mac which will be done after commit.

Reviewers: bert

Reviewed By: bert

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3391

git-svn-id: https://svn.opendtect.org/od/trunk@33741 a985855d-d384-4b45-ae33-31a410998c3c