Made re-used taskrunner reset the status bar.
Summary:
If you use a uiTaskRunner many times, the second time you run it, you will keep the message() of the previous job if the current message is "".

An example of this is the installer that says "Copying data" in the first task. In the second task (download) there is no message. As it is "", the "Copying data" will still be displayed.

Test Plan: Erronious message is not there any longer.

Reviewers: aneesh

Reviewed By: aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D3412

git-svn-id: https://svn.opendtect.org/od/trunk@33777 a985855d-d384-4b45-ae33-31a410998c3c