od4.6 : Fix for crash while session restore.
Summary:
http://bugs.dgbes.com/view.php?id=2625 ( Crash Report Mantis ID 2625 )

In uiODSceneMgr::useScenePars() function we do this...
	Scene& scn = mkNewScene();
  	if ( !scn.sovwr_->usePar(*scenepar) )
	{
	    removeScene( scn.mdiwin_->mainObject() );
	    continue;
	}

So when scn.sovwr_->usePar(*scenepar) function fails, we remove the scene.
But at this stage scene->itemmanager_ is null. So it crashes while accessing it.
So added null pointer check to avoid the crash.

scene->itemmanager_ is initialized in the initTree() function. So another way to fix this would be to call initTree before !scn.sovwr_->usePar(*scenepar) this function eg.
  
	Scene& scn = mkNewScene();
	initTree( scn, vwridx_ );
  	if ( !scn.sovwr_->usePar(*scenepar) )
	{
	    removeScene( scn.mdiwin_->mainObject() );
	    continue;
	}
In this case scene->itemmanager_ is initialized and we wouldn't need the null pointer check.

Is there a better way to fix this?

Test Plan: Tested by changing survey, session restore and adding new scene.

Reviewers: kristofer, nanne

Reviewed By: nanne

CC: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3407

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33791 a985855d-d384-4b45-ae33-31a410998c3c