Added support for uiString on osg-texts.
Summary: Enabled support for translated texts in the visualization. The saving of the actual textes in the annotation were removed from the sessions, as those texts are assumed to be restored from the system itself when starting up the scene.

Test Plan: OpendTect starts and compiles. The annotation text is as normal.

Reviewers: ding

Reviewed By: ding

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3445

git-svn-id: https://svn.opendtect.org/od/trunk@33825 a985855d-d384-4b45-ae33-31a410998c3c