od : Fixing cdash error in nightly build
Summary:
http://intranet/cdash/viewBuildError.php?buildid=32824
src/uiODMain/uiodscenemgr.cc:163: error: address of overloaded function with no contextual type information.

As there are two functions with same name removeScene, there is ambiguity while using macro mCB(obj,clss,fn).

As it is defined as 
#define mCBFn(clss,fn) ((CallBackFunction)(&clss::fn))
#define mCB(obj,clss,fn) CallBack( static_cast<clss*>(obj), mCBFn(clss,fn))

So renaming the function removeScene to doRemoveScene to avoid this ambiguity.

Another way to fix this would be to change the macro itself like 
 
#define mCBFn(clss,fn) ((CallBackFunction)((void(clss::*)(CallBacker*))&clss::fn))
#define mCB(obj,clss,fn) CallBack( static_cast<clss*>(obj), mCBFn(clss,fn))

Test Plan: compiles

Reviewers: nanne

Reviewed By: nanne

CC: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D3453

git-svn-id: https://svn.opendtect.org/od/trunk@33839 a985855d-d384-4b45-ae33-31a410998c3c