od4.6: Fix for default 3D vol in 2D Input field
Summary: If the user has defined a default 3D Volume but NO default 2D Lineset, the input fields for 2D seismics shows the dafault 3D volume at start. This has already been fixed in 4.7 in a better way. This hack is for 4.6 only.

Test Plan:
In 2D steering creation dialog,
1. If default 3D volume defined but NO dfeault 2D lineset is defined, the input field is empty at start.
2. If a default 2D lineset is defined, the input field has this default selected at start.

In both the cases, processing creates a dip-steering volume successfully.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D3447

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33851 a985855d-d384-4b45-ae33-31a410998c3c