od4.6: Bug fix for correctly creating FileName entry in .2ds file after OD5.0 has been run on the survey once.
Summary:
Run od5.0 and then create a new attribute in a lineset in OD4.6 and everything is messed up.
Current implementation uses previous entry in .2ds and increment the number in the name to create new entry in .2ds. This is not correct as now number after '^' represent GeomID.
So the check is basically to identify the format.

Test Plan:
tested by following the steps.
Run od5.0
Run 0d4.6 and create a new attribute.
Go back to od5.0 and it will convert it to its format
run od4.6 and attribute is visible.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T104

Differential Revision: https://phabricator.dgbes.com:8443/D3451

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33854 a985855d-d384-4b45-ae33-31a410998c3c