Bugfix: Mouse manipulation of draggers should be impossible when in view mode or polygon selection mode.
Summary:
Draggers were mouse manipulated when (1) rotating the scene in view mode or
(2) dragging the red polygon in polygon selection mode. The former is a little
bug solved in visdata.cc. The latter is missing functionality needed for OSG
in uivispartserv.cc. These solutions are independent of the question whether
draggers should be displayed at all in these modes.

Test Plan:
Added an inline to the scene while already in view mode. Rotating the scene
with the mouse pointer on top of the dragger no longer makes the dragger (tabs)
move. Neither when switching to polygon selection mode and start dragging the
red polygon on top of the dragger (tabs).

Reviewers: ranojay

Reviewed By: ranojay

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3455

git-svn-id: https://svn.opendtect.org/od/trunk@33856 a985855d-d384-4b45-ae33-31a410998c3c