Bugfix for data object (plane) not disappearing from scene when removed from data tree, until moving mouse over scene.
Summary:
Not every data object disappeared from scene directly on every platform when
it is removed from the data tree, because redraw-on-demand was not triggered.
It turns out that VisualObjectImpl::removeChild() is not called at every tree
object we remove, so that is why DataObjectGroup::removeObject() got a redraw
trigger as well. Does not hurt anyway.

Test Plan:
Only verified by putting breakpoints, because the actual bug did not reproduce
on my platform. May try yourself after switching back to redraw-on-demand.

Reviewers: ranojay

Reviewed By: ranojay

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3458

git-svn-id: https://svn.opendtect.org/od/trunk@33872 a985855d-d384-4b45-ae33-31a410998c3c