od: Fixed turnOn/Off bug for horizons displayed only on sections.
Summary:
When horizons are displayed only on section, turning on/off any of them was affecting on/off state for all the horizons. This was reported by Jan.

Simply restored the code as was done in od4.6, as this was the only chnage noticed there other than the osg changes.

Test Plan: Works as expected

Reviewers: ding

Reviewed By: ding

CC: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D3419

git-svn-id: https://svn.opendtect.org/od/trunk@33877 a985855d-d384-4b45-ae33-31a410998c3c