od : Fix for crash report ( Mantis ID 2678 )
Summary:
http://bugs.dgbes.com/view.php?id=2678
In this case rdmlinedlg_ is initialised as 
	rdmlinedlg_ = new uiWell2RandomLineDlg( parent(), this );
Its parent() is the active window. So when its parent is deleted, it delets all its children including rdmlinedlg_. But uiWellPartServer has no idea that its already deleted and it tries to delete it in uiWellPartServer::selectWellCoordsForRdmLine() which leads to double deletion and ultimately crash.

Hence added uiWellPartServer::rdmlnDlgDeleted( CallBacker* ) function in which we do rdmlinedlg_ = 0, so as to avoid double deletion.

Test Plan: No more crash

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3489

git-svn-id: https://svn.opendtect.org/od/trunk@33912 a985855d-d384-4b45-ae33-31a410998c3c