Made right-click menus available in view mode.
Summary:
Since objects are not "actively" pickable in viewmode, the right-click action
has been added to the passively pickable category, which furthermore includes
mouse moves. Right-click drag is occupied by the TrackBallManipulator zoom-in
and zoom-out.

Test Plan:
Tested on inline with new horizon to pick. Tried left and right button, click
and drag, view mode and act mode, in all combinations. All as expected. No
menu pops up if an active editor is the selected object in the tree, but this
neither is the case in OD4.6 and earlier.

Reviewers: nanne

Reviewed By: nanne

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3499

git-svn-id: https://svn.opendtect.org/od/trunk@33929 a985855d-d384-4b45-ae33-31a410998c3c