UiString -- wchar_t for Bug 2682
Summary:
When using wchar_t osgtext doesn't allow to set it as empty, otherwise it will be crashed. 
In od side, I added protection to avoid to add a empty wchar_t to osgText, If you think it is necessary, we can do something in osgText class in osg.

Test Plan: Compiles and Test fine

Reviewers: kristofer

Reviewed By: kristofer

Maniphest Tasks: T74

Differential Revision: https://phabricator.dgbes.com:8443/D3501

git-svn-id: https://svn.opendtect.org/od/trunk@33943 a985855d-d384-4b45-ae33-31a410998c3c