od: Added interface to set the replacement velocity
Summary:
Gave the possibily to set the replacement velocity from the time-depth model manager, instead
of having to reload the well.
Also made changes in uiWellman, to ensure that the datasets in memory correspond to the files on disk when the writing fails.

Test Plan: tested in F3_Demo and Turgeon surveys

Reviewers: bert

Reviewed By: bert

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3500

git-svn-id: https://svn.opendtect.org/od/trunk@33958 a985855d-d384-4b45-ae33-31a410998c3c